requirements.txt
# Core ML/DL frameworks
torch==2.1.0
torchvision==0.16.0
tensorflow==2.14.0
scikit-learn==1.3.2
scikit-image==0.22.0
opencv-python==4.8.1.78

# AWS SDK and SageMaker
boto3==1.34.10
sagemaker==2.199.0
awscli==1.32.10

# Data processing
pandas==2.1.4
numpy==1.26.2
Pillow==10.1.0
albumentations==1.3.1
imgaug==0.4.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Experiment tracking
mlflow==2.9.2
wandb==0.16.1

# Model interpretability
shap==0.44.0
captum==0.7.0

# API and deployment
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
moto==4.2.11  # AWS mocking

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pylint==3.0.3

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0
hydra-core==1.3.2

# Utilities
tqdm==4.66.1
click==8.1.7
requests==2.31.0
tenacity==8.2.3

# Monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
